@@ -11,14 +11,14 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/cescoferraro/spotify.git"
+    "url": "git+https://github.com/cescoferraro/craigs.git"
   },
   "author": "",
   "license": "W",
   "bugs": {
-    "url": "https://github.com/cescoferraro/spotify/issues"
+    "url": "https://github.com/cescoferraro/craigs/issues"
   },
-  "homepage": "https://github.com/cescoferraro/spotify#readme",
+  "homepage": "https://github.com/cescoferraro/craigs#readme",
   "dependencies": {
     "chart.js": "^2.5.0",
     "connected-react-router": "^4.0.0-beta.4",
@@ -26,18 +26,19 @@
     "history": "^4.5.1",
     "hoist-non-react-statics": "^1.2.0",
     "myd3": "^0.0.2",
-    "react": "15.4.2",
+    "prop-types": "^15.5.8",
+    "react": "15.5.4",
     "react-addons-css-transition-group": "^15.4.2",
     "react-auto-scale": "^1.0.3",
-    "react-chartjs-2": "^2.0.5",
-    "react-dom": "15.4.2",
+  "react-chartjs-2": "^2.0.5",
+    "react-dom": "15.5.4",
     "react-fullpage": "^0.1.16",
     "react-hammerjs": "^0.5.0",
     "react-helmet": "^4.0.0",
     "react-mixin": "^2.0.2",
     "react-modal": "^1.6.5",
     "react-redux": "5.0.2",
-    "react-redux-firebase": "^1.3.0",
+    "react-redux-firebase": "1.3.4",
     "react-redux-modal": "^0.5.2",
     "react-redux-toastr": "^4.4.6",
     "react-router": "4.0.0-alpha.6",
@@ -48,7 +49,7 @@
     "redux": "3.6.0",
     "redux-form": "^6.5.0",
     "redux-form-material-ui": "^4.1.2",
-    "redux-observable": "^0.13.0",
+    "redux-observable": "0.14.1",
     "rx-lite-dom": "7.0.3",
     "rxjs": "^5.2.0",
     "serviceworker-webpack-plugin": "^0.1.8",
@@ -83,7 +84,7 @@
     "favicons-webpack-plugin": "0.0.7",
     "fg-loadjs": "1.0.0",
     "file-loader": "0.9.0",
-    "firebase": "^3.6.8",
+    "firebase": "3.9.0",
     "flexboxgrid": "6.3.1",
     "friendly-errors-webpack-plugin": "1.1.3",
     "html-webpack-plugin": "2.28.0",
@@ -111,7 +112,6 @@
     "react-redux": "5.0.2",
     "react-router-redux": "4.0.7",
     "react-router-server": "1.1.1",
-    "react-spotify-player": "^1.0.1",
     "react-svg-loader": "2.0.0-alpha.0",
     "react-tap-event-plugin": "2.0.1",
     "reactfire": "^1.0.0",
@@ -131,8 +131,8 @@
     "webpack-chunk-hash": "^0.3.0",
     "webpack-dashboard": "0.2.1",
     "webpack-dev-middleware": "1.9.0",
-    "webpack-hot-middleware": "2.13.2",
-    "webpack-hot-server-middleware": "0.0.4",
+    "webpack-hot-middleware": "2.18.0",
+    "webpack-hot-server-middleware": "0.1.0",
     "webpack-manifest-plugin": "1.1.0",
     "webpack-visualizer-plugin": "0.1.10"
   }
