@@ -1,4 +1,5 @@
 import * as React from "react";
+import * as PropTypes from "prop-types";
 import * as hoistStatics from "hoist-non-react-statics";
 
 
@@ -9,13 +10,13 @@ export const StyleConnect = (...styles) => {
             || ComposedComponent.name || 'Component';
 
 
-        class WithStyles extends React.Component<any,any> {
+        class WithStyles extends React.Component<any, any> {
             displayName = `WithStyles(${displayName})`;
             ComposedComponent = ComposedComponent;
             removeCss;
 
             static contextTypes = {
-                insertCss: React.PropTypes.func,
+                insertCss: PropTypes.func,
             };
 
             componentWillMount() {
@@ -35,4 +36,3 @@ export const StyleConnect = (...styles) => {
         return hoistStatics(WithStyles, ComposedComponent);
     };
 };
-
