@@ -1,23 +1,22 @@
 import * as React from "react";
+import * as PropTypes from 'prop-types';
 
 
-export class UniversalStyleProvider extends React.Component <any, any> {
+export class UniversalStyleProvider extends React.Component<any, any> {
     static propTypes = {
-        children: React.PropTypes.element.isRequired,
-        onInsertCss: React.PropTypes.func.isRequired,
+        children: PropTypes.element.isRequired,
+        onInsertCss: PropTypes.func.isRequired,
     };
 
     static childContextTypes = {
-        insertCss: React.PropTypes.func.isRequired,
+        insertCss: PropTypes.func.isRequired,
     };
 
     getChildContext() {
-        return {insertCss: this.props.onInsertCss};
+        return { insertCss: this.props.onInsertCss };
     }
 
     render() {
         return React.Children.only(this.props.children);
     }
 }
-
-
